requirements.txt
Flask~=2.2.2
pymongo~=4.3.2
load_dotenv~=0.1.0